requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==24.3.0
azure-core==1.32.0
azure-identity==1.19.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.15
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
microsoft-kiota-abstractions==1.9.0
microsoft-kiota-authentication-azure==1.9.0
microsoft-kiota-http==1.9.0
microsoft-kiota-serialization-form==1.9.0
microsoft-kiota-serialization-json==1.9.0
microsoft-kiota-serialization-multipart==1.9.0
microsoft-kiota-serialization-text==1.9.0
msal==1.31.1
msal-extensions==1.2.0
msgraph-core==1.2.0
msgraph-sdk==1.17.0
multidict==6.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
std-uritemplate==2.0.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
